Finetune hyperparameters of algorithm
NN architecture.
How the hall of fame is ordered, went from purely SSE ordering, to the pareto thing SR uses.
Duplicate protection in the hall of fame for model structures (e.g. y = 2 is the same model as y = 1+1).
Trained the SR using only the predicted values at the timepoints occurring in the dataset.
Swapped the S-criterion to be mean based in stead of min based (i.e. controls are scored how well they discriminate the plausible models on average, instead of the model structures it can least well discriminate).
Played around a bit with the initial conditions and the noise sd.